refactor: change directory
