+++ src/index.tsx
@@ -8,8 +8,9 @@ import './styles/utils.css';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import Router from './router';
-import reportWebVitals from './reportWebVitals';
+
+import Router from '@/router';
+import reportWebVitals from '@/reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
