+++ src/pages/CardList/CardList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import { routes } from './routes';
+import { routes } from '@/routes';
 
 function CardList() {
   return (
@@ -41,4 +41,4 @@ function CardList() {
   );
 }
 
-export default CardList;
+export { CardList };
